// backend/src/modules/tenants/enums/tenant.enums.ts
export enum BusinessType {
  PRODUCT = 'PRODUCT',
  SERVICE = 'SERVICE',
  MANUFACTURING = 'MANUFACTURING',
  RETAIL = 'RETAIL',
  OTHER = 'OTHER',
}

export enum BusinessScale {
  STARTUP = 'STARTUP',
  SMALL = 'SMALL',
  MEDIUM = 'MEDIUM',
  LARGE = 'LARGE',
  ENTERPRISE = 'ENTERPRISE',
}

export enum TenantStatus {
  ACTIVE = 'ACTIVE',
  PENDING = 'PENDING',
  SUSPENDED = 'SUSPENDED',
  TERMINATED = 'TERMINATED',
}

export enum VerificationStatus {
  PENDING = 'PENDING',
  VERIFIED = 'VERIFIED',
  REJECTED = 'REJECTED',
}